refactor(testutil): fold AwaitN into test_agent_runner.go
Avoid a dedicated one-function file; AwaitN sits with the other generic test
helpers (CollectEvents/CollectParts/CollectTextParts).
